dicom: move EventHelper to its own module
