+++ plugins/dicom/plugin_tests/dicom_test.py
@@ -20,10 +20,9 @@
 import os
 import dicom
 import six
-import threading
 from tests import base
-from girder import events
 from girder.constants import AccessType
+from utils.event_helper import EventHelper
 
 
 def setUpModule():
@@ -33,40 +32,6 @@ def setUpModule():
 
 def tearDownModule():
     base.stopServer()
-
-
-class EventHelper(object):
-    """
-    Helper class to wait for plugin's data.process event handler to complete.
-    Usage:
-        with EventHelper('event.name') as helper:
-            self.model('upload').uploadFile(...)
-            handled = helper.wait()
-    """
-
-    def __init__(self, eventName, timeout=10):
-        self.eventName = eventName
-        self.timeout = timeout
-        self.handlerName = 'HandlerCallback'
-        self.event = threading.Event()
-
-    def wait(self):
-        """
-        Wait for the handler to complete.
-        :returns: True if the handler completes before the timeout or has
-                  already been called.
-        """
-        return self.event.wait(self.timeout)
-
-    def _callback(self, event):
-        self.event.set()
-
-    def __enter__(self):
-        events.bind(self.eventName, self.handlerName, self._callback)
-        return self
-
-    def __exit__(self, *args):
-        events.unbind(self.eventName, self.handlerName)
 
 
 class DicomTestCase(base.TestCase):
